Major autoloader fix
This is a required patch to make Rehike load on any operating system other than Windows.

Essentially, the class names were not converted to Unix forward slash path format. This isn't an issue on Windows, which handles both backslashes and forward slashes for path separation. However, it is an issue on other operating systems.