Fix guide (again)
Guide on watch was loading super inefficiently even though we had built
the necessary infrastructure to delayload it and prevent that issue. To
recap: I'm an idiot.

Now watch load speeds shouldn't be as slow.

Actually, I only found this out because optimising the guide request
elsewhere broke watch entirely, so I had to fix it this way.