Ist es over für mich? (not going well)
Here's the latest work I was able to hack up for this project. Note that
it's mostly testing code, but it would have gotten refactored into
something nicer if it were viable. Unfortunately, that doesn't seem to
be the case. I am heavily considering resetting development, given the
circumstances.

I really don't like the code quality that this seems to lock us to, and
it seems like Twig itself isn't really compatible with what we're trying
to do. Random pages here will crash because imports seemingly aren't
hoisted when working with renderBlock() in Twig.

I think the best way forward would be to return to tradition: forming
the SPF JSON itself in Twig, with only a small utilities class in PHP to
replace whatever services like that SpfPhp provided (checking if SPF is
requested, etc.). This is what Rehike originally did for SPF support,
and SpfPhp was created as a result of its bad design. Perhaps it would
best, however, to consider all views as a Twig-context thing.

Apart from that, it would certainly make cleaner code. Try to understand
whatever the fuck is going on in this commit - you cannot. It is not
possible.