Remove fnmatch polyfill
I think this was a bad pattern, since we want consistent behavior across
operating systems, don't we? The only thing encouraged by the presence
of this polyfill was the possibility of things behaving incorrectly when
a PHP extension implementation is used instead of our version.

I just decided to simplify the code, so we'll use only our version in
all cases.