Fix guide playlists (#559)
Guide playlists were broken by a recent InnerTube update. All playlists in
the user's library would appear on the guide itself, but YouTube replaced
this functionality with a link to a dedicated playlist feed. Rehike should
also explicitly support that page in the future, but guide playlists were
a much more important feature to restore.

A new class, GuidePlaylistManager, has been added. It is responsible for
managing the remote connections and caching guide playlist items. Caching
is employed for optimal performance, as to not slow down page requests by
requesting mostly-constant data every time the user changes the page. The
playlist information is retrieved from the new feed page.

In order to cope with the necessity of asynchronous code due to potentially
having to make a network request, the guide model code in general had to be
updated to be asynchronous. The insertion points are now promises, which
breaks API compatibility. This doesn't seem to be an issue, since the only
consumer of this API was HitchhikerController.

Because more information is reported on the playlists feed page, an old
feature was finally restored. Hitchhiker would display the username of the
owner channel of a saved playlist on the guide, but Kevlar removed this,
and Rehike couldn't display it previously due to it missing from the data.
The guide HTML template has been modified to restore this feature, now that
it is possible.

Signed-off-by: Isabella <kawapure@gmail.com>