Update documentation links to V3
Reprety created a website for V3. Linking to the website which can be read
under any circumstances is far nicer than linking to a Discord server that
you must join with a Discord account.

I also added the link to the documentation page "How Rehike works", and
removed a link from said page to Cosmic Cat, which is abandoned and doesn't
work anymore.

Signed-off-by: Isabella <kawapure@gmail.com>