Fix file names of log files
Log files were exclusively logging using the name "home", even though they
should have included the name of the path being visited. This is due to the
old code accessing the wrong side of a split result. It was accessing the
zeroth item of the split array after splitting by "/", but URLs typically
begin with a "/" character, so before that would just be an empty string.

I opted to clean up the algorithm entirely, ensuring that it stores a safe
file name and using PHP's proper URL parser.

Signed-off-by: Isabella <kawapure@gmail.com>