Populate new visitor info with recommendations
As part of mitigations for YouTube's recent homepage updates, this update
adds a new process to automatically select a common video (Me at the zoo)
to be watched upon generating new visitor data strings for signed out
sessions. This ensures that the homepage will have default recommendations
instead of the "watch more videos" promo.

Unfortunately, it can be a bit slow for watching a video to affect homepage
recommendations. For me, it typically takes about 15 seconds after watching
a video, which is far too long a time to block the request.

It would be preferable to add a loading screen in this case instead, though
I would rather experiment with this in another commit.

Some API changes were made in order to make this possible. Most notably, I
moved the initialisation of $yt->playerConfig over to
YtApp::bindToEverything from its previous home, HitchhikerController. The
method previously used, HitchhikerController::initPlayer, is now a no-op
and is deprecated. There are a few callers to this, so I didn't want to go
through with removing them in this commit without investigating them a bit
further, but I think it'll be alright.

I also made Network::getCurrentRequestCookie public. I think it's okay to
make this method public, as it's a method that has a primary use for making
network requests, with a descriptive name, and the implementation is very
unlikely to change nor the method be removed. This changed simply allowed a
bit of code reuse, rather than writing an identical method for a related
purpose.

Signed-off-by: Isabella <kawapure@gmail.com>