Update bug_report.yml for Rehike 0.8.1
This should have been done a long time ago, but nobody ever got around to
it, and issues continued to just use 0.8.0.

I also reordered the options so that the default option is the latest
version. A few issues were created without changing the version, but 0.7
isn't even supported anymore.

Signed-off-by: Isabella <kawapure@gmail.com>