Rename unclear names in comments AJAX templates
These variables retained the names "dv2". This referred to an early name
that Rehike used for Hitchhiker comments HTML early on: Discussion V2.

This name scheme was meant to reflect the redesigned HTML introduced in
2016, which Rehike uses, but it was abandoned in favour of a more clear
name, "comments", in 2a6dea.

For the sake of consistency, I also renamed an unrelated variable to this
commit to reflect the names in its sibiling files.

Signed-off-by: Isabella <kawapure@gmail.com>