[RHBuild] Introduce build system
This commit introduces a new build system called RehikeBuild. RHBuild is
also used as a shorthand for this new system. RehikeBuild manages building
external source code. This mostly includes external web resources like CSS
and JS, but it can also include generated code and other things too.

RehikeBuild relies on Node.js and wraps the Gulp build system. Detailed
documentation on how it works exactly will be written at a later date.

At present, the build system only supports compiling SCSS to CSS via the
sass compiler. A test project was included for testing purposes, filling
something into the empty folder at src/css/. Additional compilers (for JS
and Protobuf) will be added later.

My original plan for the system was to write something completely custom in
PHP, but it turns out that package management is actually really difficult
and we would be better off relying on a mature system with a lot of
preexisting packages like NPM. Gulp is a fairly nice build system that I
and others on the Rehike development team are already familiar with, but
the RehikeBuild code proper extends it to meet the needs we have for a
Rehike build system.

There are a couple miscellaneous changes that were made in this commit that
were made in attempts to debug during development. I think this includes
changing the sass version in package.json, but I don't remember, so I'm
keeping the change in case it actually was important.

Signed-off-by: Isabella <kawapure@gmail.com>