[RHBuild] Check in current WIP changes
I decided to do this on the master branch because RHBuild is not ready for
use yet anyway. This commit makes RHBuild once again unusable, but paves
the way for codebase restructuring that will make it easier to make build
tasks in the future.

The new architecture moves Gulp tasks to be created inside wrapper classes
which denote the input files and destination file. Support for multiple
outputs hasn't been considered yet, because I think this can be done nicely
by branching off into new tasks (once that works correctly).

This reworking allowed the earlier "rename handle" concept to be dropped;
now that the destination file names are stored in the wrapper, and the
output will be fully controlled to one single file, there is no need for
them and the architecture is much simpler as a result. This is also more
stable. We were having issues with Closure Compiler dropping our file name
during compilation and replacing it with its own output name. Now we can
just ignore Gulp tasks changing the name of the virtual file, and save it
with our intended name regardless.

I haven't yet considered how to handle the case of multiple virtual files
in the output stream from Gulp. Although build tasks can always take in
multiple files, they must output one and multiple should result in an
error.

For that matter, accessing the data from the stream is weird too. Streams
are kinda unfamiliar to me as a frontend JS developer who doesn't work with
Node that often. As I understand it, you can just asynchronously receive
any data from a stream and return it in a Promise, but maybe it would be
better to design a stream-based architectrure to begin with.

The biggest issue at the moment is that I don't know how to dynamically add
tasks while Gulp is running tasks. Callbacks have been used in the past and
present to achieve something to this effect, but ideally we would just hack
Gulp to read series data from a dynamic array which we can push to during
execution. This would work nicely with Undertaker.series (gulp.series)
conceptually, but it would be a bit weirder with parallel execution.

This is all really complicated, but if we can get this down, then we should
be able to design a very nice and fast build system for Rehike.

Signed-off-by: Isabella <kawapure@gmail.com>