[template/rehike/debugger] Remove dead code
This dead code here could've been pretty bad at that. Since
lightbox_main.twig is only imported, there was no issue, but if it were
included, it would include an additional </div> and break the markup of
whichever template included it.

Thankfully, it was just dead code, but I definitely felt it appropriate to
remove anyway.

Signed-off-by: Isabella <kawapure@gmail.com>