[i18n/en-GB,ja] Support British English, tweak Japanese
A folder for British English (en-GB) has existed since the i18n update came
out, but it's always been unfinished and hasn't been updated.

British English uniquely depends on a language inheritance system in the
CoffeeTranslation module. This feature may be used for other languages in
the future (for example, European and Latin-American Spanish), but for now,
it's only used for different English dialects.

The translation works mostly well, except for the config page. The config
page (or the LanguageApi::getAllTemplates() API) isn't aware of language
inheritance, so only unique strings from that file are retrieved and not
mixed in with records from another file. I'll leave it for Niko to fix
this issue since it's her module.

I also fixed the date/time definitions in the Japanese translation. Niko
made a small mistake when writing the comments and left the 12-hour time
depiction of the times while dropping the suffix (so 12:00 AM -> 12:00,
which is simply wrong). These comments have been corrected, and I changed
the time format slightly because I think it should always use a leading
zero rather than not.

Signed-off-by: Isabella <kawapure@gmail.com>