Correct channel tab order in some cases
Channel tabs always followed a fixed order, but Rehike had a few issues
with the corrections to restore certain tabs like about.

The about tab was hardcoded to be inserted before the search tab, so on
some channels that didn't have a search box, the about tab would be placed
at the very beginning, before the home tab even appears.

The design has been corrected so that channel tabs are now added following
a fixed order.

Example channels with no searchbox:
https://www.youtube.com/@thekiboomers
https://www.youtube.com/channel/UCzuqhhs6NWbgTzMuM09WKDQ

Signed-off-by: Isabella <kawapure@gmail.com>