Push test Nepeta code.
I had done this work a while ago, but never ended up pushing it.

Nepeta is a planned system for user extensions to Rehike.

The design at the time of making this was to use interprocess
communication to have context isolation and allow the system to be
extremely flexible. This will probably be redesigned in the future since
I couldn't figure out how to get the IPC working right, and it's totally
overboard for what Nepeta will be used for.

:33 < nepeta is going to be a purrfect extension system!