Fix C4 tabs on playlist headers
These started crashing due to the reworking, but I think I fixed that now.
Tabs previously did not get a base URL for the playlists view, which ended
up breaking a few things. I added code to guess the base URL based on the
tabs if it isn't already set by the time we get to processing the tabs, and
added additional safeguards for functions related to the tab sorting algo.

All changes should suffice for the special use of the C4 header on playlist
views.

Signed-off-by: Isabella <kawapure@gmail.com>