Tweak debug logger
I made these changes to the debug logger for local testing and thought that
they'd be valuable to check into upstream.

DebugLogger::getLogs() now returns messages in chronological order properly
instead of reverse-chronological order. This was probably the result of
some overly-defensive programming, but I don't think it was ever intended,
as even the comment itself said it is meant to go in chronological order.

I also made it avoid repeating log messages which are spammed. This really
helps with memory usage by the logger, while keeping the reporting just as
informative. This code should probably be refactored at some point, as it
was just hacked on because it was needed.

Signed-off-by: Isabella <kawapure@gmail.com>