Add log tab to Rebug
Previously, the debug logs were only visible upon a fatal error. Rebug did
not have the capability to view log messages, so the feature was not used
very often. This commit adds the ability to view debug logs in Rebug, plus
internationalisation strings to English and Japanese for the new GUI tab.

Signed-off-by: Isabella <kawapure@gmail.com>