Support on-update callbacks for config changes
This is a useful feature to have, in case a configuration change requires
some extra operations like clearing cache. To preserve the existing API
design, a new method "registerUpdateCb()" was added to the base associative
property class, which returns the same object in a chain. The only issue
is that calling such a method requires wrapping the new expression in
parentheses due to PHP syntactic limitations:

(new BoolProp(...))->registerUpdateCb(...)

But I think it's very nice, and works a bit better than extending the
constructor arguments of every property, because those are variable.

Signed-off-by: Isabella <kawapure@gmail.com>