Experiment: Tick injection for scheduling
This new experiment adds the ability to inject tick declarations into each
PHP file during runtime, which paves the path for implementing a fully-
asynchronous scheduler down the line.

PHP's tick feature allows a function to be defined which is called every
few VM instructions or so. This process itself is asynchronous and managed
by the PHP interpreter, so it is a very good base to design a proper async
scheduler.

See: https://www.php.net/manual/en/function.register-tick-function.php

I experimented with asynchronous scheduling a little bit, and it seems to
work mostly alright, except that my implementation was quick and dirty, and
didn't work properly with infrequent tick intervals.

When this experiment is enabled, the file:// protocol handler is overridden
and PHP files are rewritten before being parsed. This hooking mechanism
seems to have virtually no impact on performance, and is fully compatible
with OPCache. In fact, changing the experiment explicitly requires that the
OPCache state be discarded so that each PHP file is recompiled with the
injected changes.

Ticks themselves also seem to have little-to-no impact on performance. I
haven't profiled the code yet, but there is no tangible impact compared to
having the experiment disabled.

Asynchronous scheduling can be used to provide a better experiencing with
our existing asynchronous code. One key benefit that it offers is that
async code running in the CoffeeRequest event loop can now return to
synchronous code and continue execution there, rather than waiting on all
events to finish.

Note that a significant amount of Rehike's existing asynchronous code is
written with the consideration that calls into asynchronous code will block
execution of further synchronous code, and some things even rely on such
behaviour. More reworking will be necessary to support a true asynchronous
scheduler in Rehike.

The PHP file hooking mechanism introduced in this commit may also be used
for other purposes, such as the Nepeta extension system for user extensions
of Rehike.

Signed-off-by: Isabella <kawapure@gmail.com>