Channels4 model bakery fixes and improvements
This commit fixes... strangely enough, there are no issues on the repo for
this common issue. You know how the subscribe button is missing on the
channel page for half the users? This fixes that.

I rewrote the channels4 bakery to use an instantiated object instead of a
static class in preparation for future restructurings.

The subscribe button now appears on channel pages when the viewmodel
experiment is active. It appears to work correctly in my brief testing, but
there is at least one limitation of note: it does not get the notification
state ID, unlike the handler for classical InnerTube objects.

I also cleaned up some miscellaneous debug or otherwise leftover dead code.

Signed-off-by: Isabella <kawapure@gmail.com>