[RHBuild] Fix JS compilation; output; more!
This commit cleans up a few things and gets file output and JS compilation
working.

I finalised the decision to abandon Gulp's transform streams in favour of
promises for RehikeBuild's own code. Promises are just nicer to work with.
This does mean some boilerplate in order to synchronise the two, but it's
worth it.

JavaScript compilation now finally works. Since we already rely on directly
including the textual content of the files for bundling purposes, I just
downloaded one of the Gulp preprocessor packages and used that. This can be
replaced with a better system in the future. The Closure Compiler bundler
caused runtime errors in my testing, so it was abandoned.

Build output now works, and it's independent of Gulp! This means getting
the files to save to the correct destination is now a piece of cake.

A clone of YouTube's VFL system was added for cache-busting purposes. I
decided to make the build system generate a static cache upon receiving all
VFL files, which will help with lookup. The files are not versioned on disk
and older versions are not stored.

The codebase generally was restructured to split things away from
rehikebuild_main.js. This is done to make the code easier to maintain.

Signed-off-by: Isabella <kawapure@gmail.com>