[RHBuild] Support for building individual packages
RehikeBuild previously required rebuilding all packages. This commit allows
packages to be built individually.

There were some complications here... Firstly, the Gulp CLI utility itself
doesn't really like my approach to command-line arguments, and spaced-out
strings that don't immediately succeed a "--" or "-" pair will be
attempted-to-be-parsed-as-Gulp-task-names by the Gulp command-line parser.
I'm not sure how to work around this, but such work needs to be done.

The second issue is that VFL generation was not designed with this in mind.
The system will need to be reworked a little bit to read from the existing
file when a full compilation isn't be performed in order to prevent the
existing records from being discarded when it goes to dump the information
about the single file it ends up building. Not a very difficult issue to
solve, but I just settled for disabling the VFL cache generation altogether
for this commit.

Some further tidying up has been done, and now the Gulp CLI is really,
really becoming quite burdensome. For instance, I added support for a tiny
Visual-Studio-like build-completion message system, which shows completion
statistics about all of the packages. Unfortunately, Gulp likes to display
its own, quite-useless completion messages underneath this supposed-to-be-
final message here.

Sorry about the rambling, but that really did get annoying.

Signed-off-by: Isabella <kawapure@gmail.com>