Move Rebug JS to RHBuild package
This excludes the xhook portion for now. That will be moved in the future,
but it isn't worth the hassle right now, especially because it's only for
extra functionality and isn't initialised by the Rebug JS itself.

One part of the Rebug JS used conditional compilation, which was the
network component. This used to be completely excluded from the resulting
source code if Rebug was condensed, but I simply restructured it to not
initialise during runtime if Rebug is condensed.

Apart from that, everything seems to work perfectly fine!

Signed-off-by: Isabella <kawapure@gmail.com>