Hacky fix for modern username lookup
YouTube fucked up parsing new username URLs.

If a channel has a modern username which is the same as a legacy one, then
the legacy one is preferred by YouTube's server-side parser, and you will
visit the legacy channel.

For example:

/@Stryder7x (UCYDnJiF0_RqSjkjvjRbG1tA) ->
/user/Stryder7x (UC7O-LSXY_CWSI3AIoIsamqg)

Because this can be quite confusing, we'll just fix YouTube's fuck up in
Rehike by always using /channel/ links for /@ URLs. This will be reverted
if and when YouTube get around to fixing this parser bug.

Signed-off-by: Isabella <kawapure@gmail.com>