Fix channel banner changing (+ more)
Channel banners were broken for the following reasons:
- They made a request on the www subdomain that wasn't in the funnel list
- A bug in the funnelling process (described below)

If something goes awry during PHP execution, a warning can be triggered.
In the default PHP setup used by most Rehike users, warnings are printed by
default in the server settings and are only disabled during runtime by
Rehike's own code.

In the case of setting a channel banner, the image file sent in the request
data would trigger an "input variables exceeded 1000" warning, which would
corrupt the output JSON expected by the client application, and the upload
would be cancelled.

In order to work around this issue, I simply discard the initial output
buffer that we put first thing in the Rehike script and restart it. This
gets rid of any data that is destined to be sent to the user's browser
before the data we want (the proxied network response).

Signed-off-by: Isabella <kawapure@gmail.com>