Support new metadata on channel about
This commit adds support for displaying the channel's username, video
count, and pronouns on the channel about page. This is new information
that was only introduced to the channel page after Hitchhiker was
discontinued.

In order to maintain an authentic experience, this information is only
displayed if the user does not opt out.

Signed-off-by: Isabella <kawapure@gmail.com>