Fix some bugs mentioned in #582
I wrapped the error log file manager in an exception handler in order to
avoid throwing an uncaught exception in the general exception handler (this
obfuscates the prior exception by throwing it away).

In the future, the log manager should be restructured to check for the
availability of modules before attempting to access them.

This commit does not fix the primary bug of #582.

Signed-off-by: Isabella <kawapure@gmail.com>