Fix classic player init errors
Some changes to the player JS broke our algorithm for retrieving the N
parameter. It appears to be a combination of both intentional obfuscation
and possible compiler changes.

Of note: the "n" parameter text is no longer stored as a string literal in
the player JS, but is rather stored in the form of
`String.fromCharCode(110)`, which is almost certainly an obfuscation
measure.

They also changed how the "n" parameter function itself returns, which
broke our return parser. I'm not sure if this is a compiler artifact or an
intentional choice, but the return pattern seems to have from
    return a.join("")
to
    return Array.prototype.join.call(a, "")
where `a` is a placeholder variable name. This might be done for some weird
hack rather than obfuscation, but it's a case we need to handle
nonetheless.

This fixes #586.

Signed-off-by: Isabella <kawapure@gmail.com>