Move "narrow" utils to Rehike\Helper namespace...
This is a design pattern I picked up from some private projects I work on,
and I've grown accustomed to it.

Narrow utils are those which apply to a single aspect of the codebase (i.e.
one page, module, etc.) rather than common aspects; those which are shared
by many parts of the codebase. The latter group is referred to as broad
utilities.

It always felt a little weird to cluster narrow utilities with broad ones,
so I'm glad that this could finally be resolved in some way.

Signed-off-by: Isabella <kawapure@gmail.com>