Fix for MessyJsonParser
The parser would break when hitting a single \ character, like seen PHP
namespace paths. We recently had a commit name that broke this, because of
course it would.

It would be best to move away from using JSON for this purpose altogether,
or at least to modify the version file generation script to generate well-
formed JSON.

Signed-off-by: Isabella <kawapure@gmail.com>