Rehike 0.8.2: Change the licence
Rehike was previously licensed under the Unlicense, a very liberal, public
domain license. In this commit, we have switched to the Mozilla Public
License version 2.0, which aligns with the views of The Rehike Maintainers,
Network Neighborhood, and of the core developers therein.

The Mozilla Public License v2 (MPL 2.0) enforces that forks of Rehike must:
  - Publish their source code
  - Maintain the MPL licensing (i.e. you cannot remove the license to
    publish a closed-source release of Rehike's source code in the future).

Because we are staunchly pro-open-source, this license change reflects our
core beliefs far more than the Unlicense had.

This licence change does not affect previously published code. Therefore,
you could still technically release a closed-source product based on Rehike
version 0.8.1 or earlier, if you really wanted to... ;-;

Because of the nature of the licensing change, versions for the PlayerCore
library and Rehike itself have been incremented in their minor version by
one; PlayerCore is now version 3.2 and Rehike is now version 0.8.2.

Signed-off-by: Isabella <kawapure@gmail.com>