Fix bug with getting native name for foreign language
This wasn't working right because the names are stored in CoffeeTranslation
with underscores, but InnerTube and Rehike use hyphens.

There are interesting a few places where the InnerTube data deviates from
Unicode CLDR: Spanish (US) exists to InnerTube, but not to CLDR.

Unfortunately, as I see it, there isn't any reasonable way to retrieve
all native names for foreign languages in InnerTube anymore? WEB v2 just
gives the local name for that language (i.e. what it calls itself).

Signed-off-by: Isabella <kawapure@gmail.com>