Temporary emergency workaround for player issues
This commit introduces a temporary experiment which will move player
requests to the embedded player. About 1/5 people will get a directly-
embedded embedded player instance (iframe) instead of only the
WEB_EMBEDDED_PLAYER client being used, which should help us realise any
bugs with this workaround. It also gives people more immediate options
since it is configurable.

The current configuration option is reviewed in the video metadata in the
description of a video for the time being.

Some additional changes were done regarding player flags (we want PO tokens
when the embedded player is used primarily).

Per the advice of the V3 developer, Reprety, we are experimenting with PO
tokens first and foremost. Server ABR streaming is currently nuked for
these experiments.

Signed-off-by: Isabella <kawapure@gmail.com>