Invert weight of player temporary workaround experiments
Following some IM debugging with a user, I've come to the conclusion that
the best possible user experience (i.e. the most videos that can be played)
is provided by embedding the embed player iframe rather than using its
response. The latter approach fails to load some videos, whereas the former
approach has no problem.

Because of this, I decided to invert the weights for the time being.

Signed-off-by: Isabella <kawapure@gmail.com>