[RH Core JS] Improve event handling API.
This is a bit of progressive enhancement improving the API design of the
events API provided by the Rehike Core JS.

Instead of returning void, an object containing the specified parameters
will be returned instead. This is useful in dealing with event handlers
because it is common for arguments to be passed that are created at call
time (.bind(), anonymous callbacks).

From there, you can simply call .remove() on the returned object to remove
the event listener.

Signed-off-by: Isabella <kawapure@gmail.com>