Ponies! :3
This is an easter egg YouTube included back in 2013. I liked it so much
that I decided to include it in Rehike. When you search the name of a pony
from My Little Pony (Friendship is Magic; this doesn't account for G5 :P),
then the masthead will change its colour to the body of the searched-for
pony.

This will persist between pages until you change your search query or open
a new tab. I designed it like this so you could keep the colour when you
go to watch a video about that pony, which isn't something that happened in
2013 (I presume no SPF was an issue :P).

I also added data for Starlight Glimmer, who wasn't introduced until 2015,
and thus couldn't have existed in the 2013 easter egg :PPPPP

Signed-off-by: Isabella <kawapure@gmail.com>