Remove temporary emergency player workaround (b7c21e5)
This code has overstayed its welcome. Users are reporting good results with
video playback using the latest player with the WEB v2 API, so I don't
think that this code is necessary to keep around anymore.

In fact, I've heard more reports of video playback not working due to the
experiment using one of the embedded players.

Signed-off-by: Isabella <kawapure@gmail.com>