Fix guide model to account for new Cairo icons
Man, I hope that Cairo icons won't become a problem anywhere else. They're
quite egregious on the guide, but the guide isn't the only place where we
use icons for pattern searching, so this might be pretty bad.

This should fix part of #600, but not the whole thing. Some additional
testing with the issuer will need to be conducted, as this is untested
(I do not have the experiment) and there are reportedly other issues
happening coincidentally.

Cairo icon IDs can be found in this binary for reference:
https://www.youtube.com/s/desktop/4610dd25/jsbin/desktop_polymer.vflset/desktop_polymer.js

Signed-off-by: Isabella <kawapure@gmail.com>