Add support for new Ringo2 branding (2024)
It's easy to miss it, but YouTube have changed their branding slightly as
of 2024. The new branding, officially known as Ringo2 (yeah, that's not
a name that anyone in the YouTube modding community made up :P) refines
some aspects of the 2017 Ringo rebrand.

The Ringo rebrand of 2017 introduced the new YouTube logo and changed the
accent colour from a dark red to a bright red.

Ringo2 changed the accent colour from a bright red to a slightly more
pinkish red. There is an increased focus on red-to-pinkish-red gradients
throughout the design, which can be seen in the new favicon and loading
progress bar.

Of course, the 2017 Ringo rebrand and 2015 logo remain accessible in Rehike
as before. The configuration property was migrated to an enum in order to
support three different states, but nothing else changed.

Signed-off-by: Isabella <kawapure@gmail.com>