[Rebug] Improve handling of private properties
Previously, protected and private object properties did not show up in
Rebug dumps, which was a little confusing. This commit introduces a hack to
get them to show up properly in the global object dump.

Along with this, the Rebug GUI has been modified to explicitly state the
privacy of such objects.

Signed-off-by: Isabella <kawapure@gmail.com>