Fix erroneous history addition when logged in
This hack seems to only be intended for use when the user is logged out,
so there is no reason to run this code when the user is logged in at
all.

If this code were to be ran, which may happen in some cases where the
VISITOR_INFO1_LIVE cookie expires during an authenticated user session,
then an entry for "Me at the zoo" would be added to the user's watch
history.