Modernise watch model bakery
This code was written in Rehike's infancy, with the majority of code having
been published on 2022/06/22 in the following commit:
ffb2ceb51eb9692207899931763b86e3e4fd4295

I recommmend reading through the changelog of that commit if you wanna see
some of Rehike's early design.

Naturally, the watch page was the among the first models to be completed,
and it was never updated until now to account for us just getting better.

The watch model uniquely used a confusing split-bakery system, seemingly
with the goal of emulating virtual functions to a degree in our prevalent-
at-the-time procedural style. Since then, the norm has been to use objects
for model bakeries.

Signed-off-by: Isabella <kawapure@gmail.com>