Clean up SignInV2 implementation
It was useless to include auth user ID in the base class for all both
Google accounts and YouTube channels, since only Google accounts may have
one. The method to retrieve the auth user ID still exists in the YouTube
channel info class, but it simply retrieves the information from the owner
Google account.

BackwardsCompatibilitySessionInfoFactory (what a mouthful!) has been
changed to use a singleton pattern. The previous attempt at optimisation
had failed because a new instance of the factory was being made every time
the API was called, so the session info object was being recreated as well.
Despite that, it was surprisingly performant even without the optimisation.

Signed-off-by: Isabella <kawapure@gmail.com>