Fix InnerTube replies error handling
I intended to fix what I believed to be an error with Rehike parsing
comment responses, but then I realised that it's actually a familiar issue
with InnerTube itself. InnerTube sometimes just doesn't return anything
even remotely resembling a valid replies response, and somehow we're
expected to make something out of it.

I also believe that I've encountered this issue on the official YouTube
application on my mobile phone, so I doubt that it's anything to do with
Rehike.

Ergo, the error handling was fixed. Now an error that the frontend can
interpret is returned instead of the Rehike server throwing an exception,
meaning that the "load more replies" button will return to the pre-clicked
state with no new content. Beats the old, incorrect behaviour where it'd
just "load" forever. More accurate, too.

Signed-off-by: Isabella <kawapure@gmail.com>