Revert the memory hole (partially)
This helps in tracking who wrote which parts of the codebase, which is
generally useful.

Taniko turned out to not be a good person, and neither I, Aubrey, nor any
other fellow developers of Rehike support her actions. Restoring this
attribution was done simply to keep track of who wrote what.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>