Improve fatal error handling
Fatal errors now use the Hitchhiker pageframe via Twig if it is available.
Since they are a critical feature, they will also display even if the
Twig stack isn't yet available.

Additionally, I have made the file path use a link to open the file via
the VS Code protocol. VS Code is the preferred text editor by the majority
of Rehike maintainers, so I went with it. There is no way to open a local
file through a website other than that.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>