Work around June 2025 legacy player initialisation removal
I guess YouTube finally removed the classic way of initialising the player
which they abandoned almost a decade ago. I can't say I blame them, but it
does make me curious how much of the player is lingering legacy code that
they haven't removed. They evidently still have specific checks for WEB v1
(i.e. Hitchhiker) to enable the use of certain events that were abandoned
over 5 years ago.

Fortunately, Taniko already built up the groundwork for this workaround all
the way back in 2023 due to YouTube already starting to break stuff
(Limp Bizkit reference!) That made it significantly easier for me to make
this change today, so I must thank her for one of her very few good
contributions to the project that doesn't suck ass (CoffeeRequest and the
async framework aren't baaaaad, but CoffeeTranslation fucking sucks, also
fuck you personally, Taniko. And by the way, why the hell did you put
spaces in that function name for the patched function there when YOU should
know that that makes it harder to search for?!)

Honestly, going into this, I was expecting it to be much more of a fight.
As a matter of fact, it looks super simple. They added one new line of code
to the constructor for the player Application object (i.e. what is created
upon calling yt.player.Application.create) that will check if the argument
specifying the player context configuration, and if it is null, it will
intentionally `throw Error("b/179532961")`. My guess is that this is
explicit deprecation with the intention to remove any related code later,
because it does not seem that anything else has changed besides this. In
previous versions of the player, this line of code simply did not exist,
and so the yt.player.Application constructor accepted a null argument for
context configuration (in fact, this was an intentional part of the design
which Hitchhiker leveraged while it coexisted with the Polymer frontend;
while newer versions of Hitchhiker's base.js has some references to context
config -- probably because of sharing library code with Polymer -- it never
actually used it)

If you read that terribly-long previous paragraph, you would have noticed
that I mentioned something about deprecation. I don't think it'll be a
problem in the near future, but they might perform additional codebase
refactors that will get in our way, so it's worth consideration...
I suppose? It's not like anybody really maintains Rehike other than me, and
I barely give a damn as it is, so it's not like anybody is gonna be on the
look out.

Anyway, that was a surprisingly easy fix for this bug. I just wanted to
rant about mostly-unrelated shit regarding the project.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>